--- Happiggy-bank/Happiggy-bank/View/HomeView.swift
@@ -1,22 +1,0 @@
-//
-//  HomeView.swift
-//  Happiggy-bank
-//
-//  Created by 권은빈 on 2022/02/17.
-//
-
-import UIKit
-
-/// BottleViewController의 뷰, InitialImage 뷰를 나타낼 HomeView
-final class HomeView: UIView {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        self.frame = UIScreen.main.bounds
-        self.backgroundColor = .lightGray
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-    }
-}
